[refactor] Remove dependency tracking in Task.
Previously, we intended to mark a task of a node dependent on the tasks of its children. To do this correctly, we need to transitively search all children down the hierarchy which is inefficient. For now, we remove this idea. The programmer is allocated the responsibility for ensuring that the tasks are ordered correctly.