[feat] Change implementation of expr extraction mutation.
Previously, we first scanned the expression to get parameters. The function body is later rewritten by updating the expressions chosen as parameters. 

The new implementation does everything in one pass. 
1. Scan subexpressions to select parameters required for subexpression and to build the subexpression that will go into the body.
2. Compose subexpressions to build the parent expression. The resultant parameters are all the parameters from all the subexpressions.

This change is motivated by the dependency of parameter selection on specification building. If we cannot come up with a specification for an expression, it cannot be included in the function body and will need to be passed in as a parameter. On the other hand, specification building requires an up-to-date function body. Hence, we do everything in one pass.